+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.2] - 2025-06-13
+- improve demo sketch - add getLastError()
+- improve error handling
+- add getCRC(), getProduct(), getFactorySettings().
+- update readme.md
+- update keywords
+- minor edits
+
 ## [0.1.1] - 2025-04-07
 - update GitHub actions
 - update readme.md
@@ -13,7 +21,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - add **uint32_t lastRead()**
 - add derived class **MS5803** to be able to set address.
 - minor edits
-
 
 ## [0.1.0] - 2023-12-24
 - initial release